add client side

+++ server/database.sql
@@ -4,7 +4,7 @@ CREATE TABLE hot_dogs
 (
   id SERIAL PRIMARY KEY,
   name VARCHAR(50),
-  price INT,
+  price NUMERIC,
   description TEXT,
   image_url VARCHAR(255)
 );
